Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db>=0.1.1",
    "fastapi>=0.128.3",
    "fastlib>=0.1.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-groq>=1.1.2",
    "mongo>=0.2.0",
    "mongoengine>=0.29.1",
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "uvicorn>=0.40.0",
]
